Revert "Full overhawl of video & audio playback to make it more complete (#1432)"
This reverts commit 123a81e12d072cd95d701231176d7eb2d03b3961.
