--- .changeset/cool-grapes-hug.md
@@ -1,5 +1,0 @@
----
-'rrdom': patch
----
-
-Support `loop` in `RRMediaElement`
