--- .changeset/dirty-rules-dress.md
@@ -1,5 +1,0 @@
----
-'rrweb-snapshot': minor
----
-
-Video and Audio elements now also capture `playbackRate`, `muted`, `loop`, `volume`.
