--- .changeset/mighty-ads-worry.md
@@ -1,5 +1,0 @@
----
-'rrweb': minor
----
-
-Full overhawl of `video` and `audio` element playback. More robust and fixes lots of bugs related to pausing/playing/skipping/muting/playbackRate etc.
