--- .changeset/silver-pots-sit.md
@@ -1,5 +1,0 @@
----
-'@rrweb/types': patch
----
-
-Add `loop` to `mediaInteractionParam`
