--- .changeset/smart-geckos-cover.md
@@ -1,5 +1,0 @@
----
-'rrweb': patch
----
-
-Record `loop` on `<audio>` & `<video>` elements.
