+++ packages/rrweb-snapshot/src/rebuild.ts
@@ -348,17 +348,6 @@ function buildNode(
               break;
             default:
           }
-        } else if (
-          name === 'rr_mediaPlaybackRate' &&
-          typeof value === 'number'
-        ) {
-          (node as HTMLMediaElement).playbackRate = value;
-        } else if (name === 'rr_mediaMuted' && typeof value === 'boolean') {
-          (node as HTMLMediaElement).muted = value;
-        } else if (name === 'rr_mediaLoop' && typeof value === 'boolean') {
-          (node as HTMLMediaElement).loop = value;
-        } else if (name === 'rr_mediaVolume' && typeof value === 'number') {
-          (node as HTMLMediaElement).volume = value;
         }
       }
