+++ packages/rrweb/src/record/observer.ts
@@ -1042,7 +1042,7 @@ function initMediaInteractionObserver({
         ) {
           return;
         }
-        const { currentTime, volume, muted, playbackRate, loop } =
+        const { currentTime, volume, muted, playbackRate } =
           target as HTMLMediaElement;
         mediaInteractionCb({
           type,
@@ -1051,7 +1051,6 @@ function initMediaInteractionObserver({
           volume,
           muted,
           playbackRate,
-          loop,
         });
       }),
       sampling.media || 500,
