+++ packages/rrweb/src/replay/index.ts
@@ -47,6 +47,7 @@ import {
   ReplayerEvents,
   Handler,
   Emitter,
+  MediaInteractions,
   metaEvent,
   mutationData,
   scrollData,
@@ -80,7 +81,6 @@ import getInjectStyleRules from './styles/inject-style';
 import './styles/style.css';
 import canvasMutation from './canvas';
 import { deserializeArg } from './canvas/deserialize-args';
-import { MediaManager } from './media';
 
 const SKIP_TIME_INTERVAL = 5 * 1000;
 
@@ -141,9 +141,6 @@ export class Replayer {
 
   // Used to track StyleSheetObjects adopted on multiple document hosts.
   private styleMirror: StyleSheetMirror = new StyleSheetMirror();
-
-  // Used to track video & audio elements, and keep them in sync with general playback.
-  private mediaManager: MediaManager;
 
   private firstFullSnapshot: eventWithTime | true | null = null;
 
@@ -327,7 +324,6 @@ export class Replayer {
       this.firstFullSnapshot = null;
       this.mirror.reset();
       this.styleMirror.reset();
-      this.mediaManager.reset();
     });
 
     const timer = new Timer([], {
@@ -369,13 +365,6 @@ export class Replayer {
       this.emitter.emit(ReplayerEvents.StateChange, {
         speed: state,
       });
-    });
-    this.mediaManager = new MediaManager({
-      warn: this.warn.bind(this),
-      service: this.service,
-      speedService: this.speedService,
-      emitter: this.emitter,
-      getCurrentTime: this.getCurrentTime.bind(this),
     });
 
     // rebuild first full snapshot as the poster of the player
@@ -475,16 +464,10 @@ export class Replayer {
     };
   }
 
-  /**
-   * Get the actual time offset the player is at now compared to the first event.
-   */
   public getCurrentTime(): number {
     return this.timer.timeOffset + this.getTimeOffset();
   }
 
-  /**
-   * Get the time offset the player is at now compared to the first event, but without regard for the timer.
-   */
   public getTimeOffset(): number {
     const { baselineTime, events } = this.service.state.context;
     return baselineTime - events[0].timestamp;
@@ -544,9 +527,6 @@ export class Replayer {
    */
   public destroy() {
     this.pause();
-    this.mirror.reset();
-    this.styleMirror.reset();
-    this.mediaManager.reset();
     this.config.root.removeChild(this.wrapper);
     this.emitter.emit(ReplayerEvents.Destroy);
   }
@@ -687,10 +667,9 @@ export class Replayer {
             // Timer (requestAnimationFrame) can be faster than setTimeout(..., 1)
             this.firstFullSnapshot = true;
           }
-          this.mediaManager.reset();
-          this.styleMirror.reset();
           this.rebuildFullSnapshot(event, isSync);
           this.iframe.contentWindow?.scrollTo(event.data.initialOffset);
+          this.styleMirror.reset();
         };
         break;
       case EventType.IncrementalSnapshot:
@@ -799,14 +778,6 @@ export class Replayer {
     const collected: AppendedIframe[] = [];
     const afterAppend = (builtNode: Node, id: number) => {
       this.collectIframeAndAttachDocument(collected, builtNode);
-      if (this.mediaManager.isSupportedMediaElement(builtNode)) {
-        const { events } = this.service.state.context;
-        this.mediaManager.addMediaElements(
-          builtNode,
-          event.timestamp - events[0].timestamp,
-          this.mirror,
-        );
-      }
       for (const plugin of this.config.plugins || []) {
         if (plugin.onBuild)
           plugin.onBuild(builtNode, {
@@ -1290,14 +1261,35 @@ export class Replayer {
           return this.debugNodeNotFound(d, d.id);
         }
         const mediaEl = target as HTMLMediaElement | RRMediaElement;
-        const { events } = this.service.state.context;
-
-        this.mediaManager.mediaMutation({
-          target: mediaEl,
-          timeOffset: e.timestamp - events[0].timestamp,
-          mutation: d,
-        });
-
+        try {
+          if (d.currentTime !== undefined) {
+            mediaEl.currentTime = d.currentTime;
+          }
+          if (d.volume !== undefined) {
+            mediaEl.volume = d.volume;
+          }
+          if (d.muted !== undefined) {
+            mediaEl.muted = d.muted;
+          }
+          if (d.type === MediaInteractions.Pause) {
+            mediaEl.pause();
+          }
+          if (d.type === MediaInteractions.Play) {
+            // remove listener for 'canplay' event because play() is async and returns a promise
+            // i.e. media will evntualy start to play when data is loaded
+            // 'canplay' event fires even when currentTime attribute changes which may lead to
+            // unexpeted behavior
+            void mediaEl.play();
+          }
+          if (d.type === MediaInteractions.RateChange) {
+            mediaEl.playbackRate = d.playbackRate;
+          }
+        } catch (error) {
+          this.warn(
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/restrict-template-expressions
+            `Failed to replay media interactions: ${error.message || error}`,
+          );
+        }
         break;
       }
       case IncrementalSource.StyleSheetRule:
@@ -1374,11 +1366,6 @@ export class Replayer {
     }
   }
 
-  /**
-   * Apply the mutation to the virtual dom or the real dom.
-   * @param d - The mutation data.
-   * @param isSync - Whether the mutation should be applied synchronously (while fast-forwarding).
-   */
   private applyMutation(d: mutationData, isSync: boolean) {
     // Only apply virtual dom optimization if the fast-forward process has node mutation. Because the cost of creating a virtual dom tree and executing the diff algorithm is usually higher than directly applying other kind of events.
     if (this.config.useVirtualDom && !this.usingVirtualDom && isSync) {
