--- packages/rrweb/test/html/video.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Video</title>
-  </head>
-  <body>
-    <h1>Big Buck Bunny</h1>
-    <video muted controls>
-      <source src="assets/bunny-video.webm" type="video/webm" />
-      Your browser does not support the video element.
-    </video>
-  </body>
-  <script>
-    document.querySelector('video').currentTime = 5;
-  </script>
-</html>
